+++ src/Type/Collection/Interfaces/Validation/ValueValidatorInterface.php
@@ -2,9 +2,11 @@
 
 namespace LDL\Type\Collection\Interfaces\Validation;
 
+use LDL\Framework\Base\Contracts\ArrayFactoryInterface;
+use LDL\Framework\Base\Contracts\ToArrayInterface;
 use LDL\Type\Collection\Interfaces\CollectionInterface;
 
-interface ValueValidatorInterface
+interface ValueValidatorInterface extends ArrayFactoryInterface, ToArrayInterface
 {
     /**
      * @param CollectionInterface $collection