+++ src/Type/Collection/Types/Integer/Validator/IntegerValidator.php
@@ -2,6 +2,7 @@
 
 namespace LDL\Type\Collection\Types\Integer\Validator;
 
+use LDL\Framework\Base\Contracts\ArrayFactoryInterface;
 use LDL\Type\Collection\Interfaces\CollectionInterface;
 use LDL\Type\Collection\Interfaces\Validation\AppendItemValidatorInterface;
 use LDL\Type\Collection\Interfaces\Validation\KeyValidatorInterface;
@@ -35,4 +36,23 @@ class IntegerValidator implements AppendItemValidatorInterface, ValidatorModeInt
         throw new TypeMismatchException($msg);
     }
 
+    public function jsonSerialize() : array
+    {
+        return $this->toArray();
+    }
+
+    public static function fromArray(array $data = []): ArrayFactoryInterface
+    {
+        return new self(array_key_exists('strict', $data) ? (bool)$data['strict'] : true);
+    }
+
+    public function toArray(): array
+    {
+        return [
+            'class' => __CLASS__,
+            'options' => [
+                'strict' => $this->_isStrict
+            ]
+        ];
+    }
 }