+++ src/Type/Collection/Types/Lockable/Validator/LockingValidator.php
@@ -2,6 +2,7 @@
 
 namespace LDL\Type\Collection\Types\Lockable\Validator;
 
+use LDL\Framework\Base\Contracts\ArrayFactoryInterface;
 use LDL\Framework\Base\Contracts\LockableObjectInterface;
 use LDL\Framework\Base\Exception\LockingException;
 use LDL\Type\Collection\Interfaces\CollectionInterface;
@@ -13,7 +14,6 @@ use LDL\Type\Exception\TypeMismatchException;
 
 class LockingValidator implements AppendItemValidatorInterface, RemoveItemValidatorInterface, KeyValidatorInterface, ValueValidatorInterface
 {
-
     public function validateKey(CollectionInterface $collection, $item, $key) : void
     {
         $this->validateValue($collection, $item, $key);
@@ -35,5 +35,23 @@ class LockingValidator implements AppendItemValidatorInterface, RemoveItemValida
             $msg = 'Collection is locked, can not add or remove elements';
             throw new LockingException($msg);
         }
+    }
+
+    public function jsonSerialize() : array
+    {
+        return $this->toArray();
+    }
+
+    public static function fromArray(array $data = []): ArrayFactoryInterface
+    {
+        return new self();
+    }
+
+    public function toArray(): array
+    {
+        return [
+            'class' => __CLASS__,
+            'options' => []
+        ];
     }
 }