+++ src/Type/Collection/Types/Object/Validator/InterfaceComplianceItemValidator.php
@@ -2,6 +2,8 @@
 
 namespace LDL\Type\Collection\Types\Object\Validator;
 
+use LDL\Framework\Base\Contracts\ArrayFactoryInterface;
+use LDL\Framework\Base\Exception\ArrayFactoryException;
 use LDL\Type\Collection\Interfaces\CollectionInterface;
 use LDL\Type\Collection\Interfaces\Validation\AppendItemValidatorInterface;
 use LDL\Type\Collection\Interfaces\Validation\ValidatorModeInterface;
@@ -58,4 +60,33 @@ class InterfaceComplianceItemValidator implements AppendItemValidatorInterface, 
         throw new TypeMismatchException($msg);
     }
 
+    public function jsonSerialize() : array
+    {
+        return $this->toArray();
+    }
+
+    public static function fromArray(array $data = []): ArrayFactoryInterface
+    {
+        if(false === array_key_exists('interfaceName', $data)){
+            $msg = sprintf("Missing property 'interfaceName' in %s", __CLASS__);
+            throw new ArrayFactoryException($msg);
+        }
+
+        try{
+            return new self((string) $data['interfaceName'], array_key_exists('strict', $data) ? (bool)$data['strict'] : true);
+        }catch(\Exception $e){
+            throw new ArrayFactoryException($e->getMessage());
+        }
+    }
+
+    public function toArray(): array
+    {
+        return [
+            'class' => __CLASS__,
+            'options' => [
+                'interfaceName' => $this->interface,
+                'strict' => $this->isStrict
+            ]
+        ];
+    }
 }