+++ src/Type/Collection/Validator/MaxAmountValidator.php
@@ -2,6 +2,8 @@
 
 namespace LDL\Type\Collection\Validator;
 
+use LDL\Framework\Base\Contracts\ArrayFactoryInterface;
+use LDL\Framework\Base\Exception\ArrayFactoryException;
 use LDL\Type\Collection\Interfaces\CollectionInterface;
 use LDL\Type\Collection\Interfaces\Validation\AppendItemValidatorInterface;
 use LDL\Type\Collection\Interfaces\Validation\ValueValidatorInterface;
@@ -32,5 +34,34 @@ class MaxAmountValidator implements AppendItemValidatorInterface, ValueValidator
 
         $msg = "Items in this collection can not be more than: {$this->maxAmount}";
         throw new AmountValidatorException($msg);
+    }
+
+    public function jsonSerialize() : array
+    {
+        return $this->toArray();
+    }
+
+    public static function fromArray(array $data = []): ArrayFactoryInterface
+    {
+        if(false === array_key_exists('maxAmount', $data)){
+            $msg = sprintf("Missing property 'maxAmount' in %s", __CLASS__);
+            throw new ArrayFactoryException($msg);
+        }
+
+        try{
+            return new self((int) $data['maxAmount']);
+        }catch(\Exception $e){
+            throw new ArrayFactoryException($e->getMessage());
+        }
+    }
+
+    public function toArray(): array
+    {
+        return [
+            'class' => __CLASS__,
+            'options' => [
+                'maxAmount' => $this->maxAmount
+            ]
+        ];
     }
 }