+++ src/Type/Collection/Validator/MinNumericValidator.php
@@ -2,6 +2,8 @@
 
 namespace LDL\Type\Collection\Validator;
 
+use LDL\Framework\Base\Contracts\ArrayFactoryInterface;
+use LDL\Framework\Base\Exception\ArrayFactoryException;
 use LDL\Type\Collection\Interfaces\CollectionInterface;
 use LDL\Type\Collection\Interfaces\Validation\AppendItemValidatorInterface;
 use LDL\Type\Collection\Interfaces\Validation\KeyValidatorInterface;
@@ -60,5 +62,47 @@ class MinNumericValidator implements AppendItemValidatorInterface, ValidatorMode
 
         $msg = "Items in this collection can not be less than: {$this->value}";
         throw new NumericRangeValidatorException($msg);
+    }
+
+    /**
+     * @return array
+     */
+    public function jsonSerialize() : array
+    {
+        return $this->toArray();
+    }
+
+    /**
+     * @param array $data
+     * @return ArrayFactoryInterface
+     * @throws ArrayFactoryException
+     */
+    public static function fromArray(array $data = []): ArrayFactoryInterface
+    {
+        if(false === array_key_exists('value', $data)){
+            $msg = sprintf("Missing property 'value' in %s", __CLASS__);
+            throw new ArrayFactoryException($msg);
+        }
+
+        try{
+            return new self($data['value'], array_key_exists('strict', $data) ? (bool)$data['strict'] : true);
+        }catch(\Exception $e){
+            throw new ArrayFactoryException($e->getMessage());
+        }
+
+    }
+
+    /**
+     * @return array
+     */
+    public function toArray(): array
+    {
+        return [
+            'class' => __CLASS__,
+            'options' => [
+                'value' => $this->value,
+                'strict' => $this->_isStrict
+            ]
+        ];
     }
 }